ATO-130: Extract logic to get client from DynamoDB to AbstractValidationService
